feat: Create authorization feature

+++ src/services/notes/controllers/note-controller.js
@@ -1,4 +1,4 @@
-import {InvariantError, NotFoundError} from '../../../exceptions/index.js'
+import {AuthorizationError, InvariantError, NotFoundError} from '../../../exceptions/index.js'
 import response from '../../../utils/response.js'
 import NoteRepositories from '../repositories/note-repository.js'
 
@@ -8,7 +8,9 @@ import NoteRepositories from '../repositories/note-repository.js'
  * @returns {JSON} response
  */
 const getNotes = async (req, res) => {
-  const notes = await NoteRepositories.getNotes()
+  const {id: owner} = req.user
+
+  const notes = await NoteRepositories.getNotes(owner)
 
   return response(res, 200, 'Catatan sukses ditampilkan', {notes})
 }
@@ -20,9 +22,10 @@ const getNotes = async (req, res) => {
  */
 const createNote = async (req, res, next) => {
   const {title, body, tags} = req.validated
+  const {id: owner} = req.user
 
   // Finding created note
-  const note = await NoteRepositories.createNote({title, body, tags})
+  const note = await NoteRepositories.createNote({title, body, tags, owner})
 
   // Check if note fail to create
   if (!note) return next(new InvariantError('Catatan gagal ditambahkan'))
@@ -38,6 +41,12 @@ const createNote = async (req, res, next) => {
  */
 const getNoteById = async (req, res, next) => {
   const {id} = req.params
+  const {id: owner} = req.user
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner)
+
+  // Check if not the owner
+  if (!isOwner) return next(new AuthorizationError('Anda tidak berhak mengakses resources ini'))
 
   // Finding note by id
   const note = await NoteRepositories.getNoteById(id)
@@ -57,6 +66,12 @@ const getNoteById = async (req, res, next) => {
 const updateNoteById = async (req, res, next) => {
   const {id} = req.params
   const {title, body, tags} = req.validated
+  const {id: owner} = req.user
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner)
+
+  // Check if not the owner
+  if (!isOwner) return next(new AuthorizationError('Anda tidak berhak mengakses resources ini'))
 
   const note = await NoteRepositories.updateNoteById({id, title, body, tags})
 
@@ -74,6 +89,12 @@ const updateNoteById = async (req, res, next) => {
  */
 const deleteNoteById = async (req, res, next) => {
   const {id} = req.params
+  const {id: owner} = req.user
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner)
+
+  // Check if not the owner
+  if (!isOwner) return next(new AuthorizationError('Anda tidak berhak mengakses resources ini'))
 
   const note = await NoteRepositories.deleteNoteById(id)
